requirement.txt
anyio==3.6.1
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.1.0
pydantic==1.9.0
python-multipart==0.0.5
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
